move time utils to utils/timeUtils.ts
