+++ src/renderer/contexts/openGov/Treasury/index.tsx
@@ -7,7 +7,7 @@ import { createContext, useContext, useState } from 'react';
 import { encodeAddress } from '@polkadot/util-crypto';
 import BigNumber from 'bignumber.js';
 import { rmCommas } from '@w3ux/utils';
-import { formatBlocksToTime } from '@/model/Track';
+import { formatBlocksToTime } from '@/renderer/utils/timeUtils';
 import { chainCurrency } from '@/config/chains';
 import type { AnyData } from '@/types/misc';
 import type { TreasuryContextInterface } from './types';
